+++ java/sage/Seeker.java
@@ -48,7 +48,7 @@ import java.util.Vector;
  * VideoSetupMenu for modifying video directory, EventRouter for determining if Power
  * button turns on or off
  */
-public class Seeker implements Runnable
+public class Seeker implements Hunter
 {
   private static final String VIDEO_STORAGE = "video_storage";
   private static final String ARCHIVE_DIRECTORY = "archive_directory";
@@ -1707,7 +1707,7 @@ public class Seeker implements Runnable
     }
   }
 
-  void kick()
+  public void kick()
   {
     synchronized (this)
     {
@@ -4568,7 +4568,7 @@ if (encState.currRecord.getDuration() + (Sage.time() - encState.lastResetTime) >
     return wakeupTime;
   }
 
-  void goodbye()
+  public void goodbye()
   {
     //Set savedPartials = new HashSet(currRecordFiles);
     // NOTE: NARFLEX - 02/28/08 - This needs to be synchronized on the Seeker because the VideoFrame will
@@ -5050,7 +5050,7 @@ if (encState.currRecord.getDuration() + (Sage.time() - encState.lastResetTime) >
     }
   }
 
-  int forceChannelTune(String mmcInputName, String chanString, UIClient uiClient)
+  public int forceChannelTune(String mmcInputName, String chanString, UIClient uiClient)
   {
     if (Sage.client)
     {
@@ -5138,7 +5138,7 @@ if (encState.currRecord.getDuration() + (Sage.time() - encState.lastResetTime) >
     return null;
   }
 
-  MediaFile requestWatch(Airing watchAir, int[] errorReturn, UIClient uiClient)
+  public MediaFile requestWatch(Airing watchAir, int[] errorReturn, UIClient uiClient)
   {
     if (Sage.client)
       return NetworkClient.getSN().requestWatch(uiClient, watchAir, errorReturn);