+++ java/sage/Show.java
@@ -133,6 +133,9 @@ public final class Show extends DBObject
   public static final int IMAGE_SOURCE_BONES_GSTATIC = 0;
   public static final int IMAGE_SOURCE_FIBER_GSTATIC = 1;
 
+  // Used for localized check on isMovie() method.
+  static String movieString = Sage.rez("Movie");
+
   public static String getRoleString(int x)
   {
     if (x < 0 || x >= ROLE_NAMES.length)
@@ -202,6 +205,18 @@ public final class Show extends DBObject
 
   public String getCategory() { return (categories.length == 0) ? "" : categories[0].name; }
 
+  /**
+   * Checks case if the provided string matches the category.
+   *
+   * @param compare This is a Stringer optimized lookup that requires the incoming string to already
+   *               be all lowercase.
+   * @return <code>true</code> if the string matches the category.
+   */
+  public boolean isCategory(String compare)
+  {
+    return categories.length != 0 && categories[0].equalsIgnoreCase(compare);
+  }
+
   public String getSubCategory() { return (categories.length < 2) ? "" : categories[1].name; }
 
   public String[] getCategories()
@@ -261,7 +276,7 @@ public final class Show extends DBObject
     }
   }
 
-  public void appendExpandedRatingsString(StringBuffer rv)
+  public void appendExpandedRatingsString(StringBuilder rv)
   {
     for (int i = 0; i < ers.length; i++)
     {
@@ -273,7 +288,7 @@ public final class Show extends DBObject
 
   public String getExpandedRatingsString()
   {
-    StringBuffer rv = new StringBuffer();
+    StringBuilder rv = new StringBuilder();
     appendExpandedRatingsString(rv);
     return rv.toString();
   }
@@ -290,7 +305,7 @@ public final class Show extends DBObject
     return rv;
   }
 
-  public void appendBonusesString(StringBuffer rv)
+  public void appendBonusesString(StringBuilder rv)
   {
     for (int i = 0; i < bonuses.length; i++)
     {
@@ -302,7 +317,7 @@ public final class Show extends DBObject
 
   public String getBonusesString()
   {
-    StringBuffer rv = new StringBuffer();
+    StringBuilder rv = new StringBuilder();
     appendBonusesString(rv);
     return rv.toString();
   }
@@ -1464,7 +1479,7 @@ public final class Show extends DBObject
   {
     if (externalID.length > 1 && externalID[0] == 'M' && externalID[1] == 'V')
       return true;
-    if (categories != null && categories.length > 0 && Sage.rez("Movie").equals(categories[0].name))
+    if (categories != null && categories.length > 0 && movieString.equals(categories[0].name))
       return true;
     return false;
   }