+++ java/sage/api/FavoriteAPI.java
@@ -459,7 +459,7 @@ public class FavoriteAPI {
         if (Permissions.hasPermission(Permissions.PERMISSION_RECORDINGSCHEDULE, stack.getUIMgr()))
         {
           Carny.getInstance().setRecordingQuality(a, q);
-          Scheduler.getInstance().kick(false); // because it can change the encoders for it
+          SchedulerSelector.getInstance().kick(false); // because it can change the encoders for it
         }
         return null;
       }});
@@ -535,7 +535,7 @@ public class FavoriteAPI {
           if (mr != null && q != null && Permissions.hasPermission(Permissions.PERMISSION_RECORDINGSCHEDULE, stack.getUIMgr()))
           {
             mr.bully(q);
-            Scheduler.getInstance().kick(false);
+            SchedulerSelector.getInstance().kick(false);
           }
         }
         return null;
@@ -1003,7 +1003,7 @@ public class FavoriteAPI {
         {
           // Search through all of the Favorite objects to find the correct one
           Agent[] favs = Wizard.getInstance().getFavorites();
-          StringBuffer sbCache = new StringBuffer();
+          StringBuilder sbCache = new StringBuilder();
           for (int i = 0; i < favs.length; i++)
           {
             if (favs[i].followsTrend(a, false, sbCache))
@@ -1027,7 +1027,7 @@ public class FavoriteAPI {
         java.util.ArrayList rv = new java.util.ArrayList();
         // Search through all of the Favorite objects to find the correct one
         Agent[] favs = Wizard.getInstance().getFavorites();
-        StringBuffer sbCache = new StringBuffer();
+        StringBuilder sbCache = new StringBuilder();
         for (int i = 0; i < favs.length; i++)
         {
           if (favs[i].followsTrend(a, false, sbCache))
@@ -1177,7 +1177,7 @@ public class FavoriteAPI {
         Agent fav = (Agent) stack.pop();
         if (fav == null) return null;
         return fav.getRelatedAirings(Wizard.getInstance().getRawAccess(Wizard.AIRING_CODE,
-            Wizard.AIRINGS_BY_CT_CODE), true, false, new StringBuffer());
+            Wizard.AIRINGS_BY_CT_CODE), false, false, new StringBuilder());
       }});
     rft.put(new PredefinedJEPFunction("Favorite", "GetPotentialFavoriteAirings", new String[] { "Favorite" }, true)
     {
@@ -1195,7 +1195,7 @@ public class FavoriteAPI {
         Agent fav = (Agent) stack.pop();
         if (fav == null) return null;
         return fav.getRelatedAirings(Wizard.getInstance().getRawAccess(Wizard.AIRING_CODE,
-            Wizard.AIRINGS_BY_CT_CODE), true, false, true, new StringBuffer());
+            Wizard.AIRINGS_BY_CT_CODE), false, true, new StringBuilder());
       }});
     rft.put(new PredefinedJEPFunction("Favorite", "GetFavoriteID", new String[] { "Favorite" })
     {